feat: retry OSM loading max 3 times
A global variable passes the information if OSM download was succesful.
If not, data fetch is repeated two times.

Gulp is obivously not a suitable platform for this data loading process.
Inability to detect and handle errors properly and modify action flow based on other actions
makes the system  very hard to adjust, error prone and cryptic.